refactor email action handler
